requirements.txt
pandas
streamlit
altair
garminconnect
requests
python-dateutil
google-genai
plotly
streamlit-authenticator